sql database added, corrected after first review
