+++ src/main/java/ru/yandex/practicum/filmorate/service/GenreService.java
@@ -1,0 +1,23 @@
+package ru.yandex.practicum.filmorate.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.model.Genre;
+import ru.yandex.practicum.filmorate.storage.GenreStorage;
+
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+public class GenreService {
+    private final GenreStorage genreStorage;
+
+    public List<Genre> findAll() {
+        return genreStorage.findAll();
+    }
+
+    public Genre getGenre(Integer id)  {
+        return genreStorage.getGenre(id);
+    }
+
+}
