+++ src/main/java/ru/yandex/practicum/filmorate/service/LikeService.java
@@ -1,0 +1,26 @@
+package ru.yandex.practicum.filmorate.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.storage.LikeStorage;
+
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+public class LikeService {
+    private final LikeStorage likeStorage;
+
+    public void addLike(Integer filmID, Integer userID) {
+        likeStorage.addLike(filmID, userID);
+    }
+
+    public void deleteLike(Integer filmID, Integer userID) {
+        likeStorage.deleteLike(filmID, userID);
+    }
+
+    public List<Film> getMostLikedFilms(Integer count) {
+        return likeStorage.getLikedUsersID(count);
+    }
+}
